Free wireguard_device after netif is removed
The wireguard_device stored as the netif state pointer must remain valid until
after the netif is removed.  During removal, packets may be sent to disconnect
TCP sockets and such requiring a valid device for the wireguardif_output
function.

Move the device free from the wireguardif_shutdown function into a new
wireguardif_fini function which is called after the netif_remove.  Add a device
check in wireguardif_output as a precaution.

Make an explicit call to clear the IPv4 address of the Wireguard interface when
disconnecting.  The call mimics the behavior in netif_remove however it is
performed before the Wireguard peers are shutdown.  As a result, the TCP reset
packets can be sent.